Create feed module
